Don't prevent managers from applying their role-less manageds
When creating a new managed on the fly, this new person has no roles at
all. In this moment it is required that the manager can create a
participation for this role-less managed, regardless of whether the
managed is allowed to see the event or not.
Extending this, if the manager aborts the application process right
after creating their new managed, but before creating the participation,
they need to be able to come back later and add their managed to the
event.

This was implemented in 4b2a53b, but due to merge conflicts, a feature
spec was accidentally restored later. This commit here removes the
feature spec again, and cleans up some more.